==> notebooks/6_SMOD_caclulation.ipynb <==
## 6_SMOD_caclulation
### Adds SMOD definition to the labelled data set
### Make sure the detailed level segregation from the 8_rural_urban_json_segregation notebook is present in the data curation bucket
### Make sure the SMOD definitions are exact values from the JSON file mentioned above. Configure the SMOD_mapper if needed

### Initial configuration
#### To start working with this particular notebook, you need to provide necessary credential and settings
#### Below is an template of configuration, which is necessary prepare aside of this notebook and copy & paste all content in triple quotes to the next cell's input field
    """
    {
    "COS_ENDPOINT_URL": "s3.private.eu-de.cloud-object-storage.appdomain.cloud",
    "COS_AUTH_ENDPOINT_URL": "https://iam.cloud.ibm.com/oidc/token",
    "COS_APIKEY": "xxx",
    "DATA_CURATION_BUCKET": "xxx",
    "SMOD_DEFINITION_JSON": "Kenya_segmentation_2025.json"
    }
    """


# Read notebook configuration
import getpass
import json

config_str = getpass.getpass('Enter your prepared config: ')
config = json.loads(config_str)

# Import necessary libraries
import json
import geopandas as gpd
import pandas as pd
import shapely
from tqdm import tqdm
from collections import Counter
from IPython.display import clear_output
from botocore.client import Config
import ibm_boto3
import io


# init S3 client in order to upload data to the curation bucket
cos_client = ibm_boto3.client(service_name='s3',
                              ibm_api_key_id=config["COS_APIKEY"],
                              ibm_auth_endpoint=config["COS_AUTH_ENDPOINT_URL"],
                              config=Config(signature_version='oauth'),
                              endpoint_url=config["COS_ENDPOINT_URL"])

labelled_data_parquet = 'all_labelled_data.parquet'
labelled_data_SMOD_parquet = 'all_labelled_data_SMOD.parquet'
smod_polygons_json = config["SMOD_DEFINITION_JSON"]
curation_bucket = config["DATA_CURATION_BUCKET"]

# Fetch the labelled data set and the SMOD polygons
if type(curation_bucket) == str:

    streaming_body = cos_client.get_object(Bucket=curation_bucket, Key=smod_polygons_json)['Body']
    print("Downloading to local storage :  " + smod_polygons_json)
    with io.FileIO(smod_polygons_json, 'w') as file:
        for i in io.BytesIO(streaming_body.read()):
            file.write(i)
    
    streaming_body2 = cos_client.get_object(Bucket=curation_bucket, Key=labelled_data_parquet)['Body']
    print("Downloading to local storage :  " + labelled_data_parquet)
    with io.FileIO(labelled_data_parquet, 'w') as file:
        for i in io.BytesIO(streaming_body2.read()):
            file.write(i)

main_df = gpd.read_parquet(labelled_data_parquet)

segmentation = gpd.read_file(smod_polygons_json)
segmentation

# Change the definition if needed
SMOD_mapper = {
    # 'LOW_DENSITY_RURAL': 1,
    'RURAL_CLUSTER': 2,
    'SUBURBAN_PERI_URBAN': 3,
    'SEMI_DENSE_URBAN': 4,
    'DENSE_URBAN': 5,
    'URBAN_CENTER': 6,
}


SMOD_items = [k for k in SMOD_mapper.keys()]


main_df['SMOD_name'] = ['LOW_DENSITY_RURAL' for _ in range(len(main_df))]
main_df['SMOD_id'] = [1 for _ in range(len(main_df))]

for smod_name, smod_id in SMOD_mapper.items():
    
    smod_polygons = segmentation[segmentation.seg_type == smod_name].geometry
    
    for poly in tqdm(smod_polygons, total=len(smod_polygons), desc=f'Processing: {smod_name}'):
        
        min_lon, min_lat, max_lon, max_lat = poly.bounds
        
        buildings_in_bbox = main_df[
                                    (main_df.latitude >= min_lat) &
                                    (main_df.latitude <= max_lat) &
                                    (main_df.longitude >= min_lon) &
                                    (main_df.longitude <= max_lon)
                                    ].copy()
        
        if len(buildings_in_bbox) > 0:
            
            buildings_in_bbox['buildings_in_polygon'] = [poly.contains(shapely.Point(row.longitude, row.latitude)) for row in buildings_in_bbox.itertuples()]
            builings_in_polygon = buildings_in_bbox[buildings_in_bbox['buildings_in_polygon'] == True]
            
            main_df.loc[main_df.index.isin(list(builings_in_polygon.index)), 'SMOD_name'] = smod_name
            main_df.loc[main_df.index.isin(list(builings_in_polygon.index)), 'SMOD_id'] = smod_id


main_df[main_df.SMOD_name == '']

Counter(main_df.SMOD_name)

main_df = main_df.drop_duplicates(subset='geometry')
main_df

main_df.to_parquet(labelled_data_SMOD_parquet)

# optionaly upload file to the bucket
if type(curation_bucket) == str:
        
    try:
        cos_client.upload_file(
            Filename=labelled_data_SMOD_parquet,
            Bucket=curation_bucket,
            Key=labelled_data_SMOD_parquet,
            ExtraArgs={'ContentDisposition': 'attachment'}
        )
           
        print(f'File {labelled_data_SMOD_parquet} successfully uploaded to the COS {curation_bucket} bucket')
    except Exception as e:
        print(f"\033[91mFailed upload file to the bucket {curation_bucket}. Error: {e}")
